desktop_client/requirements.txt
python-dotenv
websockets
pystray
pillow
pytest
